+++ src/observable/IntervalObservable.ts
@@ -1,6 +1,6 @@
 import { Subscriber } from '../Subscriber';
 import { isNumeric } from '../util/isNumeric';
-import { Scheduler } from '../Scheduler';
+import { IScheduler } from '../Scheduler';
 import { Observable } from '../Observable';
 import { async } from '../scheduler/async';
 
@@ -12,7 +12,7 @@ import { async } from '../scheduler/async';
 export class IntervalObservable extends Observable<number> {
   /**
    * Creates an Observable that emits sequential numbers every specified
-   * interval of time, on a specified Scheduler.
+   * interval of time, on a specified IScheduler.
    *
    * <span class="informal">Emits incremental numbers periodically in time.
    * </span>
@@ -23,8 +23,8 @@ export class IntervalObservable extends Observable<number> {
    * ascending integers, with a constant interval of time of your choosing
    * between those emissions. The first emission is not sent immediately, but
    * only after the first period has passed. By default, this operator uses the
-   * `async` Scheduler to provide a notion of time, but you may pass any
-   * Scheduler to it.
+   * `async` IScheduler to provide a notion of time, but you may pass any
+   * IScheduler to it.
    *
    * @example <caption>Emits ascending numbers, one every second (1000ms)</caption>
    * var numbers = Rx.Observable.interval(1000);
@@ -35,7 +35,7 @@ export class IntervalObservable extends Observable<number> {
    *
    * @param {number} [period=0] The interval size in milliseconds (by default)
    * or the time unit determined by the scheduler's clock.
-   * @param {Scheduler} [scheduler=async] The Scheduler to use for scheduling
+   * @param {Scheduler} [scheduler=async] The IScheduler to use for scheduling
    * the emission of values, and providing a notion of "time".
    * @return {Observable} An Observable that emits a sequential number each time
    * interval.
@@ -44,7 +44,7 @@ export class IntervalObservable extends Observable<number> {
    * @owner Observable
    */
   static create(period: number = 0,
-                scheduler: Scheduler = async): Observable<number> {
+                scheduler: IScheduler = async): Observable<number> {
     return new IntervalObservable(period, scheduler);
   }
 
@@ -63,7 +63,7 @@ export class IntervalObservable extends Observable<number> {
   }
 
   constructor(private period: number = 0,
-              private scheduler: Scheduler = async) {
+              private scheduler: IScheduler = async) {
     super();
     if (!isNumeric(period) || period < 0) {
       this.period = 0;