+++ src/operator/startWith.ts
@@ -1,4 +1,4 @@
-import { Scheduler } from '../Scheduler';
+import { IScheduler } from '../Scheduler';
 import { Observable } from '../Observable';
 import { ArrayObservable } from '../observable/ArrayObservable';
 import { ScalarObservable } from '../observable/ScalarObservable';
@@ -7,13 +7,13 @@ import { concatStatic } from './concat';
 import { isScheduler } from '../util/isScheduler';
 
 /* tslint:disable:max-line-length */
-export function startWith<T>(this: Observable<T>, v1: T, scheduler?: Scheduler): Observable<T>;
-export function startWith<T>(this: Observable<T>, v1: T, v2: T, scheduler?: Scheduler): Observable<T>;
-export function startWith<T>(this: Observable<T>, v1: T, v2: T, v3: T, scheduler?: Scheduler): Observable<T>;
-export function startWith<T>(this: Observable<T>, v1: T, v2: T, v3: T, v4: T, scheduler?: Scheduler): Observable<T>;
-export function startWith<T>(this: Observable<T>, v1: T, v2: T, v3: T, v4: T, v5: T, scheduler?: Scheduler): Observable<T>;
-export function startWith<T>(this: Observable<T>, v1: T, v2: T, v3: T, v4: T, v5: T, v6: T, scheduler?: Scheduler): Observable<T>;
-export function startWith<T>(this: Observable<T>, ...array: Array<T | Scheduler>): Observable<T>;
+export function startWith<T>(this: Observable<T>, v1: T, scheduler?: IScheduler): Observable<T>;
+export function startWith<T>(this: Observable<T>, v1: T, v2: T, scheduler?: IScheduler): Observable<T>;
+export function startWith<T>(this: Observable<T>, v1: T, v2: T, v3: T, scheduler?: IScheduler): Observable<T>;
+export function startWith<T>(this: Observable<T>, v1: T, v2: T, v3: T, v4: T, scheduler?: IScheduler): Observable<T>;
+export function startWith<T>(this: Observable<T>, v1: T, v2: T, v3: T, v4: T, v5: T, scheduler?: IScheduler): Observable<T>;
+export function startWith<T>(this: Observable<T>, v1: T, v2: T, v3: T, v4: T, v5: T, v6: T, scheduler?: IScheduler): Observable<T>;
+export function startWith<T>(this: Observable<T>, ...array: Array<T | IScheduler>): Observable<T>;
 /* tslint:disable:max-line-length */
 
 /**
@@ -28,8 +28,8 @@ export function startWith<T>(this: Observable<T>, ...array: Array<T | Scheduler>
  * @method startWith
  * @owner Observable
  */
-export function startWith<T>(this: Observable<T>, ...array: Array<T | Scheduler>): Observable<T> {
-  let scheduler = <Scheduler>array[array.length - 1];
+export function startWith<T>(this: Observable<T>, ...array: Array<T | IScheduler>): Observable<T> {
+  let scheduler = <IScheduler>array[array.length - 1];
   if (isScheduler(scheduler)) {
     array.pop();
   } else {